SET UP Add javascript script to post api/scripts route
