+++ server/index.js
@@ -3,6 +3,7 @@ const app = express()
 const path = require('path')
 const cors = require('cors')
 const bodyParser = require('body-parser')
+const findMethod = require('./src/script.js').findMethod;
 
 // Serve static files from the React frontend app
 app.use(express.static(path.join(__dirname, '/client/public')))
@@ -30,6 +31,8 @@ app.post('/api/scripts', function (req, res) {
     exec(`python ./server/src/script.py ${input} ${output}`, function (error, stdout, stderr) {
       return res.send(stdout)
     })
+  } else if (language === 'javascript'){
+    return res.send(findMethod(input, output))
   }
 });
